Reduce time of Markowitz search.
Maintain the minimum column and row count; start Markowitz search from
this count rather than from 1. (In fact, the old version had an
asymptotic bug.)

Adapt list methods to maintain the index of the first non-empty list.
Also add method lu_list_move() for convenience.

For consistency we now always call lu_list_move() in the pivot
operation, even if the nonzero count of the updated column or row does
not change. This was done inconsistently between the different pivot
eliminations in the old version. The change can result in different
tie-breaking, so that results are not bit-compatible to the old version.