Change handling rank deficiencies.
When a column is dropped from the active submatrix in lu_pivot(), do not
increment rankdef. Instead, Markowitz search returns pivot_col >= 0 but
pivot_row < 0 if an empty column exists in the active submatrix. The
column is then eliminated without pivot operation in
lu_factorize_bump(). This leads to a better structure of the code
because rank and rankdef are increased at the same place.